Resolving and binding
Note: had to spend spend quite a bit of time debugging a bug caused by the expressions being records instead of classes like in the example, which made them compared by value by default, which broke the local scope dictionary that uses expressions as keys. Had to give it a ReferenceEqualityCompared to fix it.