Break and continue statements
For the purpose of `continue` in for loops, i originally attempted to keep the for->while rewrite logic in the parser, and attach the increment statement to the `continue` ast node. The problem is that it breaks local variable resolution, as now the same instance of the incrementer expression node exists in two different scopes (in the continue node, and nested within one block in the body for normal flow). And since the concept of for loops is leaked to the interpreter anyway when it would have to evaluate the incrementer on `continue`, this info is attached directly to the loop so it's always in one place. Appropriately `ForStatement` becomes `LoopStatement`.